ui: combining robot account hooks into a single file (PROJQUAY-0000)
