+++ web/src/hooks/useRobotAccounts.ts
@@ -9,6 +9,12 @@ import {
   IRobotTeam,
   IRobot,
   createRobotAccount,
+  bulkDeleteRepoPermsForRobot,
+  bulkUpdateRepoPermsForRobot,
+  fetchRobotPermissionsForNamespace,
+  IRepoPerm,
+  fetchRobotAccountToken,
+  regenerateRobotToken,
 } from 'src/resources/RobotsResource';
 import {updateTeamForRobot} from 'src/resources/TeamResources';
 import {useOrganizations} from 'src/hooks/UseOrganizations';
@@ -181,4 +187,249 @@ interface createNewRobotForNamespaceParams {
   reposToUpdate: IRobotRepoPerms[];
   selectedTeams: IRobotTeam[];
   robotDefaultPerm: string;
+}
+
+export function useRobotPermissions({orgName, robName, onSuccess, onError}) {
+  const [namespace, setNamespace] = useState(orgName);
+  const [robotName, setRobotName] = useState(robName);
+  const [page, setPage] = useState(1);
+  const [perPage, setPerPage] = useState(10);
+
+  const {usernames} = useOrganizations();
+  const isUser = usernames.includes(orgName);
+
+  const {
+    data: robotPermissions,
+    isLoading: loading,
+    error,
+  } = useQuery(
+    ['Namespace', namespace, 'robot', robotName, 'permissions'],
+    ({signal}) =>
+      fetchRobotPermissionsForNamespace(namespace, robotName, isUser, signal),
+    {
+      enabled: true,
+      placeholderData: [],
+      onSuccess: (result) => {
+        onSuccess(result);
+      },
+      onError: (err) => {
+        onError(err);
+      },
+    },
+  );
+
+  const queryClient = useQueryClient();
+  const deleteRepoPermsMutator = useMutation(
+    async (repoNames: string[]) => {
+      await bulkDeleteRepoPermsForRobot(namespace, robName, repoNames, isUser);
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries(['Namespace', namespace, 'robots']);
+        queryClient.invalidateQueries([
+          'Namespace',
+          namespace,
+          'robot',
+          robotName,
+          'permissions',
+        ]);
+        onSuccess();
+      },
+      onError: (err) => {
+        onError(err);
+      },
+    },
+  );
+
+  const updateRepoPermsMutator = useMutation(
+    async (repoPerms: IRepoPerm[]) => {
+      return await bulkUpdateRepoPermsForRobot(namespace, robotName, repoPerms);
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries(['Namespace', namespace, 'robots']);
+        queryClient.invalidateQueries([
+          'Namespace',
+          namespace,
+          'robot',
+          robotName,
+          'permissions',
+        ]);
+        onSuccess();
+      },
+      onError: (err) => {
+        onError(err);
+      },
+    },
+  );
+
+  return {
+    result: robotPermissions,
+    loading: loading,
+    error,
+    setPage,
+    setPerPage,
+    page,
+    perPage,
+    setNamespace,
+    namespace,
+    setRobotName,
+
+    // Mutations
+    updateRepoPerms: async (repoPerms: IRepoPerm[]) =>
+      updateRepoPermsMutator.mutate(repoPerms),
+    deleteRepoPerms: async (repoNames: string[]) =>
+      deleteRepoPermsMutator.mutate(repoNames),
+  };
+}
+
+interface bulkUpdateRepoPermsParams {
+  robotName: string;
+  repoPerms: IRepoPerm[];
+}
+
+interface bulkDeleteRepoPermsParams {
+  robotName: string;
+  repoNames: string[];
+}
+
+export function useRobotRepoPermissions({namespace, onSuccess, onError}) {
+  const queryClient = useQueryClient();
+  const [robotName, setRobotName] = useState('');
+
+  const {usernames} = useOrganizations();
+  const isUser = usernames.includes(namespace);
+
+  const deleteRepoPermsMutator = useMutation(
+    async ({robotName, repoNames}: bulkDeleteRepoPermsParams) => {
+      setRobotName(robotName);
+      return await bulkDeleteRepoPermsForRobot(
+        namespace,
+        robotName,
+        repoNames,
+        isUser,
+      );
+    },
+    {
+      onSuccess: (result) => {
+        queryClient.invalidateQueries(['Namespace', namespace, 'robots']);
+        queryClient.invalidateQueries([
+          'Namespace',
+          namespace,
+          'robot',
+          `${namespace}+${result.robotname}`,
+          'permissions',
+        ]);
+        onSuccess();
+      },
+      onError: (err) => {
+        onError(err);
+      },
+    },
+  );
+
+  const updateRepoPermsMutator = useMutation(
+    async ({robotName, repoPerms}: bulkUpdateRepoPermsParams) => {
+      setRobotName(robotName);
+      return await bulkUpdateRepoPermsForRobot(
+        namespace,
+        robotName,
+        repoPerms,
+        isUser,
+      );
+    },
+    {
+      onSuccess: (result) => {
+        queryClient.invalidateQueries(['Namespace', namespace, 'robots']);
+        queryClient.invalidateQueries([
+          'Namespace',
+          namespace,
+          'robot',
+          `${namespace}+${result.robotname}`,
+          'permissions',
+        ]);
+        onSuccess();
+      },
+      onError: (err) => {
+        onError(err);
+      },
+    },
+  );
+
+  return {
+    deleteRepoPerms: async (params: bulkDeleteRepoPermsParams) =>
+      deleteRepoPermsMutator.mutate(params),
+    updateRepoPerms: async (params: bulkUpdateRepoPermsParams) =>
+      updateRepoPermsMutator.mutate(params),
+  };
+}
+
+interface bulkUpdateRepoPermsParams {
+  robotName: string;
+  repoPerms: IRepoPerm[];
+}
+
+interface bulkDeleteRepoPermsParams {
+  robotName: string;
+  repoNames: string[];
+}
+
+export function useRobotToken({orgName, robName, onSuccess, onError}) {
+  const {usernames} = useOrganizations();
+  const isUserOrganization = usernames.includes(orgName);
+  const [namespace, setNamespace] = useState(orgName);
+  const [robotName, setRobotName] = useState(robName);
+  const [page, setPage] = useState(1);
+  const [perPage, setPerPage] = useState(10);
+  const queryClient = useQueryClient();
+
+  const {data: robotAccountToken, isLoading: loading} = useQuery(
+    ['Namespace', namespace, 'robot', robotName, 'token'],
+    ({signal}) =>
+      fetchRobotAccountToken(namespace, robotName, isUserOrganization, signal),
+    {
+      enabled: true,
+      placeholderData: {},
+      onSuccess: (result) => {
+        onSuccess(result);
+      },
+      onError: (err) => {
+        onError(err);
+      },
+    },
+  );
+
+  const regenerateRobotTokenMutator = useMutation(
+    async ({namespace, robotName}: regenerateRobotTokenParams) => {
+      return regenerateRobotToken(namespace, robotName, isUserOrganization);
+    },
+    {
+      onSuccess: (result) => {
+        queryClient.invalidateQueries([
+          'Namespace',
+          namespace,
+          'robot',
+          robotName,
+          'token',
+        ]);
+        onSuccess(result);
+      },
+      onError: (err) => {
+        onError(err);
+      },
+    },
+  );
+
+  return {
+    robotAccountToken: robotAccountToken,
+    loading: loading,
+
+    regenerateRobotToken: async (regenerateRobotTokenParams) =>
+      regenerateRobotTokenMutator.mutate(regenerateRobotTokenParams),
+  };
+}
+
+interface regenerateRobotTokenParams {
+  namespace: string;
+  robotName: string;
 }
